+++ CHANGES.rst
@@ -62,6 +62,7 @@
 - Improve cookbook and setup docs
 - Faster algorithm for defining object and background regions
 - Restore armsgs -d functionality
+- Finished cython to python conversions, but more testing needed
 
 0.7 (2017-02-07)
 ----------------